clean up and repackaging of messagebroker package
