+++ src/main/kotlin/hexagonalarchitecture/adapter/inbound/messagebroker/configuration/pubsub/PubSubConfiguration.kt
@@ -1,16 +1,11 @@
-package hexagonalarchitecture.adapter.inbound.messagebroker.pubsub.configuration
+package hexagonalarchitecture.adapter.inbound.messagebroker.configuration.pubsub
 
 import com.google.cloud.spring.pubsub.core.PubSubTemplate
-import com.google.cloud.spring.pubsub.integration.AckMode
 import com.google.cloud.spring.pubsub.integration.inbound.PubSubInboundChannelAdapter
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.integration.channel.DirectChannel
-import org.springframework.integration.channel.QueueChannel
-import org.springframework.integration.dsl.IntegrationFlow
-import org.springframework.integration.dsl.integrationFlow
-import org.springframework.messaging.Message
 import org.springframework.messaging.MessageChannel
 
 @Configuration
@@ -26,18 +21,6 @@ class PubSubConfiguration {
     ): PubSubInboundChannelAdapter =
         PubSubInboundChannelAdapter(pubSubTemplate, "json-topic-sub").apply {
             outputChannel = inboundChannel
-            ackMode = AckMode.MANUAL
+            //ackMode = AckMode.MANUAL
         }
-
-//    @Bean
-//    fun intFlow(pubsubInputChannel: MessageChannel): IntegrationFlow {
-//        return integrationFlow {
-//            route<Message<*>, String>({ it.headers["eventEntity"].toString().lowercase() }) {
-//                suffix("InputChannel")
-//            }
-//        }
-//    }
-//
-//    @Bean
-//    fun personInputChannel(): MessageChannel = DirectChannel()
 }
